requirements.txt
openai==0.28
matplotlib
seaborn
plotly==5.18.0
scipy==1.10.1
python-dotenv==1.0.0
streamlit
pandas
numpy


